Add first attempt at solution for part 1 of day 16
Gah, this hasn't worked. I was attempting to write a branch and bound
solver, but my traversal structure is taking too long. I sort of
defaulted to a best-first traversal, but with this problem that seems to
be breadth-first for most of the solution. We're not bounding out enough
branches.

It's also ended up a bit over-complicated I think. I'm think I'm going
to wind back and start again.